rework iframe-to-iframe communication to be mediated via host
instead of having the iframes talk to each other over a MessageChannel,
now they only talk to the host which forwards messages between them,
essentially acting as a mesasage broker.
this lets us skip expensive reinits we had to do before --
each frame gets a stable channel to communicate on,
and doesn't need to care about what's going on in the other frame.